% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedRep.R
\name{ExpectedRepresentation}
\alias{ExpectedRepresentation}
\title{Compute the expected degree of representation for any group in a political body}
\usage{
ExpectedRepresentation(PopShares, BodyN)
}
\arguments{
\item{PopShares}{A numeric vector containing the group-level population proportions.}

\item{BodyN}{A positive integer denoting the size of the political body in question.}

\item{a=-0.5, b=1}{Parameters controlling the affine transformation for how the representation measure is summarized.
More precisely, \code{a} and \code{b} control how the expected L1 deviation of the population shares from the body shares
are re-weighted (the expected L1 deviation is multiplied by \code{a}; \code{b} is as an additive re-scaling term). By default,
\code{a=-0.5} and \code{b=1} so that the expected Rose Index of Proportionality is returned.}
}
\value{
The expected degree of representation (a scalar).
}
\description{
Finds the degree of expected representation for any group in a political body under a random sampling model (Gerring, Jerzak and Oncel, 2022+).
}
\examples{

ExpectedRep <- ExpectedRepresentation(PopShares = c(1/3, 2/3,1/3),
                                      BodyN = 50)

print( ExpectedRep )

}
