% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarRep.R
\name{ResidualRepresentation}
\alias{ResidualRepresentation}
\title{Compute the amount of representation not explained by a random sampling model.}
\usage{
ResidualRepresentation(PopShares, BodyN)
}
\arguments{
\item{PopShares}{A numeric vector containing the group-level population proportions.}

\item{BodyN}{A positive integer denoting the size of the political body in question.}

\item{a=-0.5, b=1}{Parameters controlling the affine transformation for how the representation measure is summarized.
More precisely, \code{a} and \code{b} control how the L1 deviation of the population shares from the body shares
are re-weighted (the expected L1 deviation is multiplied by \code{a}; \code{b} is as an additive re-scaling term). By default,
\code{a=-0.5} and \code{b=1} so that the expected Rose Index of Proportionality is returned.}
}
\value{
A summary of the amount of representation not explained by a random sampling model.
More precisely, this function returns the
the residual standard deviation when using the expected degree of representation to predict observed
representatoin under a random sampling model.
}
\description{
Finds the residual standard deviation when using the expected representation for any group
in a political body to predict observed representation (Gerring, Jerzak and Oncel, 2022+).
}
\examples{

ResidualRep <- ResidualRepresentation(PopShares = c(1/3, 2/3,1/3),
                                      BodyN = 50)

print( ResidualRep )

}
